pyproject.toml
[project]
name = "wildfire-spotting"
version = "0.1.0"
description = ""
authors = [
    {name = "Ridwan Yusuf",email = "ridwan.yusuf@whiteshield.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "protobuf (>=7.34.1,<8.0.0)",
    "pyogrio (>=0.12.1,<0.13.0)",
    "pandas (>=2.2)",
    "geopandas (>=1.1.3,<2.0.0)",
    "xarray (>=2026.2.0,<2027.0.0)",
    "cdsapi (>=0.7.7,<0.8.0)",
    "earthengine-api (>=1.7.18,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "netcdf4 (>=1.7.4,<2.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "cartopy (>=0.25.0,<0.26.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "fastparquet (>=2026.3.0,<2027.0.0)",
    "mlflow (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
black = "^26.3.1"
ruff = "^0.15.7"
pytest = "^9.0.2"

